[#28] Replace Results with panics for internal errors
According to the rust book (https://doc.rust-lang.org/book/ch09-03-to-panic-or-not-to-panic.html), you should panic when internal assumptions about values or logic are broken in ways that the user cannot recover from. This describes 90% of the cases in which we were using results (for internal logic bugs). This commit replaces those all with panics. This cleans up a lot of stuff by removing results from a ton of functions. This should make it easier to trace down logic errors and also just generally keep return types cleaner.