[#38] Fix panic when edge buffer is only 1 tile wide
In very small worlds, the edge buffer can be only 1 tile wide. In those cases, there was a panic caused by a bug in NumRange. We would create a range where min==max, and when trying to map values, we would divide by (max-min), which resulted in a NaN elevation. The fix is to define an edge case when min==max, we always map to the min of the target range. Also added a test for the failing world config.