[#35] Fix duplicated runoff bug
There was a bug in runoff generation, where runoff that overflowed from one basin into another never got removed from the original, so it magically added more runoff to the system. I reorganized some code, and added a bunch of sanity check asserts to the runoff code to make sure we're not accidentally introducing runoff when we shouldn't be.