Fix lakes at edge of world
Previously we would get small lakes forming at the edge of the world on smaller worlds (e.g. radius=50). This was because we treated non-existent tiles as being at sea level. Clusters of tiles at the edge of small worlds can be below sea-level, but not large enough to become oceans. That meant they would collect runoff instead of dumping into the void.